Added levels and adjustments

+++ src/org/academiadecodigo/codezillas/Game.java
@@ -5,56 +5,75 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 
 public class Game {
 
+    private static int levelCounter = 1;
+
     public static void main(String[] args) {
 
         Platform[] platforms;
-        Canvas canvas;
+        Picture background;
         Player player;
         DonkeyZilla enemy;
         Projectile[] projectiles;
         CollisionDetector collisionDetector;
-        
 
+        if(levelCounter == 1) {
 
+            Object[] gameObjects = LevelFactory.level1();
+            for (Object object : gameObjects) {
+                if (object instanceof Picture) {
+                    background = (Picture) object;
+                }
 
+                if (object instanceof Platform[]) {
+                    platforms = (Platform[]) object;
+                }
 
-        ////PROJECTILE TESTS
+                if (object instanceof DonkeyZilla) {
+                    enemy = (DonkeyZilla) object;
+                }
 
-        projectiles = new Projectile[15];
-        int positionPlacer = 0;
+                if (object instanceof Projectile[]) {
+                    projectiles = (Projectile[]) object;
+                }
 
-        for (int i = 0; i<projectiles.length; i++) {
-
-            projectiles[i] = new Projectile(i + positionPlacer,0);
-            projectiles[i].draw();
-            positionPlacer += 1;
-        }
-
-        while(true) {
-
-            projectiles[0].fallingDown();
-
-            try {
-                Thread.sleep(60);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
+                if (object instanceof Player) {
+                    player = (Player) object;
+                }
             }
 
-            if (projectiles[0].getPos().getRow() == 800) {
-                projectiles[0].delete();
-                projectiles[0].setPos(0, 0);
-                projectiles[0].resetFireball();
-                projectiles[0].draw();
+
+            ////PROJECTILE TESTS
+
+            projectiles = new Projectile[15];
+            int positionPlacer = 0;
+
+            for (int i = 0; i < projectiles.length; i++) {
+
+                projectiles[i] = new Projectile(i + positionPlacer, 0);
+                projectiles[i].draw();
+                positionPlacer += 1;
+            }
+
+            for (int i = 0; i < 10; i++) {
+
+                projectiles[0].fallingDown();
+
+                try {
+                    Thread.sleep(60);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+
+                if (projectiles[0].getPos().getRow() == 800) {
+                    projectiles[0].delete();
+                    projectiles[0].setPos(0, 0);
+                    projectiles[0].resetFireball();
+                    projectiles[0].draw();
+                }
+
             }
         }
-
-        ///END PROJECTILE TESTS
-
-
-
-
 
 
     }
-
 }
